Change server.py to use client.py for individual connections
This removes a bunch of duplication of code and puts all the meat in the
client.py code. Now the server just listens on it's port and spawns
clients when a connection is made. This will also allow the addition of
code to handle serial ports to be put in the client.py code.